vdt_api/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-prometheus==2.3.1
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.1
prometheus_client==0.20.0
psycopg2==2.9.9
PyJWT==2.8.0
redis==5.0.5
sqlparse==0.5.0
